Exceptions were added for: wrong price, wrong id (room or reservation), empty list request, wrong user's command, wrong room type input, wrong booking attempt (no such room or it's already has been booked)

+++ src/main/java/org/example/controller/AdminController.java
@@ -1,8 +1,13 @@
 package org.example.controller;
 
+import org.example.exception.EmptyListException;
+import org.example.exception.InvalidCommandException;
+import org.example.exception.InvalidPriceException;
+import org.example.exception.InvalidRoomTypeException;
 import org.example.model.Message;
 import org.example.model.Reservation;
 import org.example.model.Type;
+import org.example.model.Workspace;
 import org.example.service.ReservationService;
 import org.example.service.WorkspaceService;
 
@@ -22,15 +27,35 @@ public class AdminController {
     }
 
     public void addWorkspace(final String type, final int price) {
-        if (isValidType(type) && price > 0) {
-            workspaceService.addWorkspace(Type.valueOf(type.toUpperCase()), price);
-            System.out.println(Message.SUCCESSFUL);
+        if (!isValidType(type)) {
+            throw new InvalidRoomTypeException("Invalid roomtype: " + type);
         } else {
-            System.out.println(Message.NOT_SUCCESSFUL);
+            if (price < 0) {
+                throw new InvalidPriceException("Price can not be below zero: " + price);
+            } else {
+                workspaceService.addWorkspace(Type.valueOf(type.toUpperCase()), price);
+                System.out.println(Message.SUCCESSFUL);
+            }
         }
     }
 
+    public Workspace getWorkspaceById(final int id) {
+        return workspaceService.getAllWorkspaces()
+                .stream()
+                .filter(n -> n.getId() == id)
+                .findFirst()
+                .orElse(null);
+    }
+
+    public List<Workspace> getAllWorkspaces(){
+        return workspaceService.getAllWorkspaces();
+    }
+
     public void removeWorkspace(final int id) {
+        List<Workspace> workspaceList = workspaceService.getAllWorkspaces();
+        if (workspaceList.isEmpty() || !workspaceList.contains(getWorkspaceById(id))) {
+            throw new InvalidCommandException("Wrong id or no working places");
+        }
         workspaceService.removeWorkspace(
                 workspaceService.getAllWorkspaces()
                         .stream()
@@ -40,7 +65,12 @@ public class AdminController {
     }
 
     public List<Reservation> getAllReservations() {
-        return reservationService.getAllReservations();
+        List<Reservation> reservationList = reservationService.getAllReservations();
+        if (reservationList.isEmpty()) {
+            throw new EmptyListException("Reservation list is empty");
+        } else {
+            return reservationList;
+        }
     }
 
     private boolean isValidType(final String type) {
